remove comments & move utils to lib
