+++ src/app/layout-metadata.ts
@@ -1,9 +1,8 @@
 import type { Metadata } from 'next';
 
-const title = 'Retro Funding 5: Voting';
-const description = 'Voting is now live for Retro Funding 5: OP Stack';
-// needs to update to op-rf5
-const url = 'https://op-rf4-voting-ui-production.up.railway.app/';
+const title = 'Retro Funding 6: Voting';
+const description = 'Voting is now live for Retro Funding 6: OP Stack';
+const url = 'http://round6.optimism.io/';
 
 export const metadata: Metadata = {
   title,
