+++ src/components/ballot/feedback-form.tsx
@@ -127,10 +127,6 @@ function createQuestions(
         />
       ),
     },
-    // {
-    //   title: `Did you observe any behavior among your fellow badgeholders that could be considered one of the following (select all that apply)?`,
-    //   children: <Behaviors />,
-    // },
     {
       title:
         'How worried are you about detrimental behavior among badgeholders influencing the allocation of Retro Funding in this round?',
@@ -203,128 +199,6 @@ function createQuestions(
         />
       ),
     },
-    // {
-    //   title: "To what extent do you trust the opinions of other badgeholders?",
-    //   children: (
-    //     <SelectForm
-    //       key="trust"
-    //       name="trust"
-    //       options={Array(7)
-    //         .fill(0)
-    //         .map((_, index) => ({
-    //           label: `${index + 1} ${
-    //             index === 0
-    //               ? "(very low trust)"
-    //               : index === 3
-    //               ? "(some trust)"
-    //               : index === 6
-    //               ? "(very high trust)"
-    //               : ""
-    //           }`,
-    //           value: String(index + 1),
-    //         }))}
-    //     />
-    //   ),
-    // },
-    // QUESTION: Are we keeping the rest of these questions?
-    // {
-    //   title:
-    //     "How would you rate your knowledge on the arguments for or against deducting external funding (e.g., VC funding, Optimism grants, or other grants) from Retro Funding rewards?",
-    //   children: (
-    //     <SelectForm
-    //       key="knowledge"
-    //       name="knowledge"
-    //       options={Array(7)
-    //         .fill(0)
-    //         .map((_, index) => ({
-    //           label: `${index + 1} ${
-    //             index === 0
-    //               ? "(very low knowledge)"
-    //               : index === 3
-    //               ? "(some knowledge)"
-    //               : index === 6
-    //               ? "(very high knowledge)"
-    //               : ""
-    //           }`,
-    //           value: String(index + 1),
-    //         }))}
-    //     />
-    //   ),
-    // },
-
-    // {
-    //   title:
-    //     "How satisfied do you feel with the definition of profit, compared to round 4?",
-    //   description:
-    //     "Definition: Impact = Award in OP, no past funding, grants, or revenue are deducted from Projects Retro Funding rewards",
-    //   children: (
-    //     <SelectForm
-    //       key="satisfaction"
-    //       name="satisfaction"
-    //       options={Array(7)
-    //         .fill(0)
-    //         .map((_, index) => ({
-    //           label: `${index + 1} ${
-    //             index === 0
-    //               ? "(not satisfied)"
-    //               : index === 3
-    //               ? "(somewhat satisfied)"
-    //               : index === 6
-    //               ? "(very satisfied)"
-    //               : ""
-    //           }`,
-    //           value: String(index + 1),
-    //         }))}
-    //     />
-    //   ),
-    // },
-    // {
-    //   title: "How understandable and clear did you find the impact metrics?",
-    //   children: (
-    //     <SelectForm
-    //       key="understandable"
-    //       name="understandable"
-    //       options={Array(7)
-    //         .fill(0)
-    //         .map((_, index) => ({
-    //           label: `${index + 1} ${
-    //             index === 0
-    //               ? "(not understandable)"
-    //               : index === 3
-    //               ? "(somewhat understandable)"
-    //               : index === 6
-    //               ? "(very clear)"
-    //               : ""
-    //           }`,
-    //           value: String(index + 1),
-    //         }))}
-    //     />
-    //   ),
-    // },
-    // {
-    //   title:
-    //     "How confident are you in the data sources and trust signals used for the impact metrics?",
-    //   children: (
-    //     <SelectForm
-    //       key="confidence_data"
-    //       name="confidence_data"
-    //       options={Array(7)
-    //         .fill(0)
-    //         .map((_, index) => ({
-    //           label: `${index + 1} ${
-    //             index === 0
-    //               ? "(not confident)"
-    //               : index === 3
-    //               ? "(somewhat confident)"
-    //               : index === 6
-    //               ? "(very confident)"
-    //               : ""
-    //           }`,
-    //           value: String(index + 1),
-    //         }))}
-    //     />
-    //   ),
-    // },
   ];
 }
