+++ src/lib/projectUtils.ts
